Code block: rework the instruction formatting functions
A new module, `instructions`, was created to house the functions.
4 ways in which code blocks can be incorrect are considered:

1. The code is not within a code block at all
2. Incorrect characters are used for back ticks
3. A language is not specified
4. A language is specified incorrectly

Splitting it up into these 4 cases allows for more specific and relevant
instructions to be shown to users. If a message has both incorrect back
ticks and an issue with the language specifier, the instructions for
fixing both issues are combined.

The instructions show a generic code example rather than using the
original code from the message. This circumvents any ambiguities when
parsing their message and trying to fix it. The escaped code block also
failed to preserve indentation. This was a problem because some users
would copy it anyway and end up with poorly formatted code. By using
a simple example that doesn't rely on indentation, it makes it clear
the example is not meant to be copied. Finally, the new examples are
shorter and thus make the embed not as giant.